Add support for MyListener call getting updates to service TXT records (2nd attempt) (#166)
Add support for MyListener call getting updates to service TXT records

At the moment, the implementation supports notification to the  ServiceListener class for additions and removals of service, but for service updates to the TXT record, the client must poll the ServiceInfo class. This draft PR provides a mechanism to have a callback on the ServiceListener class be invoked when the TXT record changes.
# Conflicts:
#	zeroconf.py